+++ includes/modules/brain/TranscriptAnalyzer.hpp
