+++ src/modules/brain/IntentEngine.cpp
@@ -1,4 +1,4 @@
-#include "IntentEngine.hpp"
+#include "modules/brain/IntentEngine.hpp"
 
 IntentEngine::IntentEngine()
 {
